Woosh du main, gestion des cas où ça ne parse pas

+++ etape-3/test/serverTest.c
@@ -20,13 +20,17 @@
 #include "response.h"
 
 _Token *getElement(_Token *root, char *name, string *s) {
-	_Token *t = searchTree(root, name);
-	int size;
-	char *str = (t != NULL)?getElementValue(t->node, &size):NULL;
-	if(s != NULL) {
-		s->base = str;
-		s->length = size;
+	_Token *t = NULL;
+	if(root != NULL) {
+		t = searchTree(root, name);
+		int size;
+		char *str = (t != NULL)?getElementValue(t->node, &size):NULL;
+		if(s != NULL) {
+			s->base = str;
+			s->length = size;
+		}
 	}
+	
 
 	return t;
 }
@@ -34,7 +38,6 @@ _Token *getElement(_Token *root, char *name, string *s) {
 int main(int argc, char *argv[])
 {
 	message *requete;
-	int res;
 	while ( 1 ) {
 		int connection_state = CLOSE_CONNECTION;
 
@@ -45,19 +48,18 @@ int main(int argc, char *argv[])
 		printf("#########################################\nDemande recue depuis le client %d\n",requete->clientId);
 		printf("Client [%d] [%s:%d]\n",requete->clientId,inet_ntoa(requete->clientAddress->sin_addr),htons(requete->clientAddress->sin_port));
 		printf("Contenu de la demande %.*s\n\n",requete->len,requete->buf);
-		if (res=parseur(requete->buf,requete->len)) {
-			_Token *tok, *r, *root, *t1 = NULL, *t2 = NULL, *t3 = NULL, *t4 = NULL, *t5 = NULL, *t6 = NULL, *t7 = NULL;
 
-			//writeDirectClient(requete->clientId,REPONSE,strlen(REPONSE));
+		_Token *message_token, *root, *t1 = NULL, *t2 = NULL, *t3 = NULL, *t4 = NULL, *t5 = NULL, *t6 = NULL, *t7 = NULL;
+		int code, headers_uniques, version_ok, path_len;
+		string *type_mime = NULL, path = { NULL, 0 }, method = { NULL, 0 }, body = { NULL, 0 }, content_length = { NULL, 0 }, http_version = { NULL, 0 }, connection_option = { NULL, 0 }, request_target = { NULL, 0 }, host = { NULL, 0 };
+
+		printf("\n\n===================================== DÉBUT DU TRAITEMENT =====================================\n\n");
+		if (parseur(requete->buf,requete->len)) { //Si le parseur génère bien un arbre
 			root=getRootTree();
-			r=searchTree(root,"HTTP_message");
-			tok = r;
-			while (tok) {
+			message_token=searchTree(root,"HTTP_message");
 
-				//========================= ZONE DE TEST =============================
-				printf("\n\n===================================== DÉBUT DE LA ZONE DE TEST =====================================\n\n");
-				int code, headers_uniques, version_ok, path_len;
-				string *type_mime = NULL, path = { NULL, 0 }, method, body, content_length, http_version, connection_option, request_target, host;
+			if (message_token != NULL) { //Si il y a bien un message
+				
 				t1 = getElement(root, "method", &method);
 
 				//Unicité des headers
@@ -96,44 +98,37 @@ int main(int argc, char *argv[])
 				// Disponibilité de la ressource
 				if(code == 200) {
 					t6 = getElement(root, "request_target", &request_target);
-
-					//if(host.base != NULL) printf("host = \"%s\" l=%d\n", host.base, host.length);
-					//else printf("Pas de host\n");
 					char *p = isAvailable(&request_target, &host, &path_len);
 					path.base = p;
 					path.length = path_len;
 
 					if (p == NULL) code = 404;
-					else type_mime = get_mime_type(&path);
+					else type_mime = get_mime_type(&path); //Type mime
 				}
-
-				printf("Code final = %d\n", code);
-
-				t7 = getElement(root, "connection_option", &connection_option);
-				connection_state = get_connection_state(&http_version, &connection_option);
-				send_response(&method, code, &path, type_mime, connection_state, requete);
-
-				printf("\n\n===================================== FIN DE LA ZONE DE TEST =====================================\n\n");
-				//=====================================================================
-
-				//int l;
-				//char *s = getElementValue(tok->node, &l);
-				//if (s != NULL) writeDirectClient(requete->clientId,s,l);
-
-				if(t1 != NULL) purgeElement(&t1);
-				if(t2 != NULL) purgeElement(&t2);
-				if(t3 != NULL) purgeElement(&t3);
-				if(t4 != NULL) purgeElement(&t4);
-				if(t5 != NULL) purgeElement(&t5);
-				if(t6 != NULL) purgeElement(&t6);
-				if(t7 != NULL) purgeElement(&t7);
-				tok=tok->next;
+			} else {
+				code = 400;
 			}
-			purgeElement(&r);
-			purgeTree(root);
 		} else {
-			//writeDirectClient(requete->clientId,ERROR,strlen(ERROR));
+			code = 400;
 		}
+
+		printf("Code final = %d\n", code);
+
+		t7 = getElement(root, "connection_option", &connection_option);
+		connection_state = get_connection_state(&http_version, &connection_option); //Gestion de la connection
+		send_response(&method, code, &path, type_mime, connection_state, requete); //Envoi de la réponse
+
+		if(message_token != NULL) purgeElement(&message_token);
+		if(t1 != NULL) purgeElement(&t1);
+		if(t2 != NULL) purgeElement(&t2);
+		if(t3 != NULL) purgeElement(&t3);
+		if(t4 != NULL) purgeElement(&t4);
+		if(t5 != NULL) purgeElement(&t5);
+		if(t6 != NULL) purgeElement(&t6);
+		if(t7 != NULL) purgeElement(&t7);
+		if(root != NULL) purgeTree(root);
+
+		printf("\n\n===================================== FIN DU TRAITEMENT =====================================\n\n");
 
 		if ((connection_state & CLOSE_CONNECTION) == CLOSE_CONNECTION) {
 			requestShutdownSocket(requete->clientId);
